fix: make android compile in the new repo sturcture
