+++ android/gradle.properties
@@ -1,5 +1,5 @@
 TurboWebview_kotlinVersion=1.7.0
-TurboWebview_minSdkVersion=21
+TurboWebview_minSdkVersion=24
 TurboWebview_targetSdkVersion=31
 TurboWebview_compileSdkVersion=31
 TurboWebview_ndkversion=21.4.7075529
