+++ android/src/main/java/com/reactnativeturbowebview/RNSession.kt
@@ -1,4 +1,4 @@
-package com.hotwirerndemo
+package com.reactnativeturbowebview
 
 import android.content.Context
 import android.util.Log
