+++ android/src/main/java/com/reactnativeturbowebview/RNVisitableView.kt
@@ -1,4 +1,4 @@
-package com.hotwirerndemo
+package com.reactnativeturbowebview
 
 import android.content.Context
 import android.graphics.Bitmap
